Added guaranteed value and delimitation flows that also emits markers
